Move numpy hs from watershed into npstats
